refactor(webapp): move node Tooltip from BlockProducers to InformationCard/Nodes
