+++ webapp/src/routes/BlockProducers/index.js
@@ -6,8 +6,6 @@ import LinearProgress from '@mui/material/LinearProgress'
 import Pagination from '@mui/material/Pagination'
 
 import SearchBar from '../../components/SearchBar'
-import Tooltip from '../../components/Tooltip'
-import NodeCard from '../../components/NodeCard'
 import InformationCard from '../../components/InformationCard'
 import useBlockProducerState from '../../hooks/customHooks/useBlockProducerState'
 import NoResults from '../../components/NoResults'
@@ -49,8 +47,6 @@ const Producers = () => {
   const classes = useStyles()
   const [
     {
-      anchorEl,
-      current,
       filters,
       chips,
       items,
@@ -60,22 +56,13 @@ const Producers = () => {
       pagination,
     },
     {
-      handlePopoverClose,
       handleOnSearch,
-      handlePopoverOpen,
       handleOnPageChange,
     },
   ] = useBlockProducerState()
 
   return (
-    <div>
-      <Tooltip
-        anchorEl={anchorEl}
-        open={anchorEl !== null}
-        onClose={handlePopoverClose}
-      >
-        <NodeCard node={current?.node} producer={current?.producer} />
-      </Tooltip>
+    <>
       <div className={classes.searchWrapper}>
         <SearchBar
           filters={filters}
@@ -95,7 +82,6 @@ const Producers = () => {
                   type="entity"
                   producer={{ ...producer, missedBlocks }}
                   rank={producer.rank}
-                  onNodeClick={handlePopoverOpen}
                 />
               </div>
             ))
@@ -111,7 +97,7 @@ const Producers = () => {
         onPageChange={handleOnPageChange}
         loading={loading}
       />
-    </div>
+    </>
   )
 }
